create component light
