+++ Engine/Module/ModuleLight.h
@@ -2,11 +2,22 @@
 #define _MODULELIGHT_H_
 #include "Module.h"
 
+class ComponentLight;
+
 class ModuleLight : public Module
 {
 public:
 	ModuleLight() = default;
-	~ModuleLight() = default;
+	~ModuleLight();
+
+
+	bool CleanUp() override;
+
+	ComponentLight* CreateComponentLight();
+	void RemoveComponentLight(ComponentLight* light_to_remove);
+
+public:
+	std::vector<ComponentLight*> lights;
 };
 
 #endif // !_MODULELIGHT_H_
