+++ Engine/Module/ModuleTexture.h
@@ -12,7 +12,6 @@ struct SDL_Renderer;
 struct SDL_Rect;
 
 class ComponentMaterial;
-class ComponentLight;
 class Texture;
 class MaterialImporter;
 
@@ -27,9 +26,6 @@ public:
 
 	ComponentMaterial* CreateComponentMaterial();
 	void RemoveComponentMaterial(ComponentMaterial* material_to_remove);
-
-	ComponentLight* CreateComponentLight();
-	void RemoveComponentLight(ComponentLight* light_to_remove);
 	
 	std::shared_ptr<Texture> LoadTexture(const char* texture_path);
 	GLuint  LoadCubemap(std::vector<std::string> faces) const;
@@ -38,7 +34,6 @@ public:
 public:
 	GLuint checkerboard_texture_id;
 	std::vector<ComponentMaterial*> materials;
-	std::vector<ComponentLight*> lights;
 
 };
